shared/python/pyproject.toml
[project]
name = "skillbeam-shared"
version = "0.1.0"
description = "Shared domain models and utilities for SkillBeam services"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6,<3.0",
  "pydantic-settings>=2.2,<3.0",
  "sqlalchemy>=2.0,<3.0",
  "psycopg2-binary>=2.9,<3.0",
  "alembic>=1.13,<2.0",
  "boto3>=1.34,<2.0",
  "python-jose[cryptography]>=3.3,<4.0",
  "celery>=5.3,<6.0",
  "redis>=5.0,<6.0",
  "pypdf>=4.0,<5.0",
  "pdfminer.six>=20231228,<20300000",
  "python-docx>=1.1,<2.0",
  "python-pptx>=0.6.23,<1.0",
  "reportlab>=4.1,<5.0",
  "openpyxl>=3.1,<4.0",
  "httpx>=0.27,<1.0",
  "youtube-transcript-api>=0.6,<1.0",
  "yt-dlp>=2025.1.0",
  "opentelemetry-api>=1.24,<2.0",
  "opentelemetry-sdk>=1.24,<2.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*"]
